requirements.txt
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
coverage==5.5
cycler==0.10.0
Django==2.2.9
django-cors-headers==3.7.0
django-debug-toolbar==3.2
django-extensions==3.1.1
djangorestframework==3.12.4
groups==0.1
idna==2.10
iniconfig==1.1.1
kiwisolver==1.3.1
matplotlib==3.4.1
numpy==1.20.2
packaging==20.9
Pillow==8.2.0
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
pytest-django==3.8.0
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
response==0.5.0
scipy==1.6.3
six==1.15.0
sorl-thumbnail==12.7.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
